Terminado layout e request
- TODO: Passar informações da cidade pra tela de detalhes;
- TODO: Terminar algoritmo que converte previsões para resumo em dias;
- TODO: Alimentar layout com informações;

+++ app/src/main/java/com/devmaker/luizmatias/weatherapp_devmaker/Previsao.java
@@ -6,14 +6,14 @@ package com.devmaker.luizmatias.weatherapp_devmaker;
 
 public class Previsao {
 
-    private long timestamp; //timestamp do horário da previsao
+    private int timestamp; //timestamp do horário da previsao
     private Cidade cidade; //objeto da cidade com a previsão do tempo
 
-    public long getTimestamp() {
+    public int getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(long timestamp) {
+    public void setTimestamp(int timestamp) {
         this.timestamp = timestamp;
     }
